extract shared/testing package for MSW, Playwright, fixtures [IEX-2469]
