+++ mocks/mock-server/capital.ts
@@ -20,8 +20,8 @@ import {
 } from '../mock-data';
 import { endpoints } from '../../endpoints/endpoints';
 import { DefaultBodyType, http, HttpResponse, StrictRequest } from 'msw';
-import { calculateGrant, delay } from './utils/utils';
-import { getHandlerCallback, mocksFactory } from './utils/mocksHandlerFactory';
+import { calculateGrant } from './utils/utils';
+import { delay, getHandlerCallback, mocksFactory } from '@integration-components/testing/msw';
 import { paths as CapitalGrantOfferPaths } from '../../src/types/api/resources/CapitalGrantOffersResource';
 import { paths as CapitalGrantsPaths } from '../../src/types/api/resources/CapitalGrantsResource';
 import { paths as CapitalMissingActionsPaths } from '../../src/types/api/resources/CapitalMissingActionsResource';
