+++ mocks/mock-server/disputes.ts
@@ -1,5 +1,5 @@
 import { http, HttpResponse, PathParams, StrictResponse } from 'msw';
-import { compareDates, delay, getPaginationLinks } from './utils/utils';
+import { compareDates, delay, getPaginationLinks } from '@integration-components/testing/msw';
 import { endpoints } from '../../endpoints/endpoints';
 import { DISPUTE_PAYMENT_SCHEMES } from '../../src/components/utils/disputes/constants';
 import { IDisputeDetail, IDisputeListItem, IDisputeStatusGroup, IDisputeListResponse } from '../../src/types/api/models/disputes';
