+++ mocks/mock-server/transactions.ts
@@ -29,7 +29,7 @@ import Localization from '../../src/core/Localization';
 import { endpoints } from '../../endpoints/endpoints';
 import { delay as mswDelay, http, HttpResponse, PathParams } from 'msw';
 import { parsePaymentMethodType } from '../../src/components/external/TransactionsOverview/components/utils';
-import { compareDates, computeHash, delay, getPaginationLinks } from './utils/utils';
+import { compareDates, computeHash, delay, getPaginationLinks } from '@integration-components/testing/msw';
 import { clamp, getMappedValue } from '../../src/utils';
 import { setupBasicResponse } from './setup';
