+++ mocks/mock-server/utils/utils.ts
@@ -1,54 +1,6 @@
-import { delay as mswDelay, DelayMode, HttpHandler } from 'msw';
 import { IGrantOfferResponseDTO } from '../../../src';
 import uuid from '../../../src/utils/random/uuid';
 
-export const compareDates = (dateString1: string, dateString2: string, operator: 'ge' | 'le') => {
-    const date1 = new Date(dateString1);
-    const date2 = new Date(dateString2);
-
-    switch (operator) {
-        case 'ge':
-            return date1 >= date2;
-        case 'le':
-            return date1 <= date2;
-    }
-};
-
-/**
- * Hash function based on {@link https://theartincode.stanis.me/008-djb2/ djb2} algorithm
- */
-export function computeHash(...strings: string[]) {
-    const hash = strings.reduce((hash, string) => {
-        let i = string.length;
-        while (i) hash = (hash * 33) ^ string.charCodeAt(--i);
-        return hash;
-    }, 5381);
-    return (hash >>> 0).toString(16).padStart(8, '0');
-}
-
-export async function delay(duration?: DelayMode | number): Promise<void> {
-    // Ensure there is no response delay in tests.
-    return Number(process.env.TEST_ENV) === 1 ? mswDelay(0) : mswDelay(duration);
-}
-
-export function getMockHandlers(mocks: HttpHandler[][]): HttpHandler[] {
-    const handlers = [] as HttpHandler[];
-    mocks.forEach(mocks => handlers.push(...mocks));
-    return handlers;
-}
-
-export const getPaginationLinks = (cursor: number, limit: number, totalLength: number) => {
-    const potentialNextCursor = cursor + limit;
-    const nextCursor = potentialNextCursor < totalLength ? potentialNextCursor : undefined;
-
-    const potentialPrevCursor = cursor - limit;
-    const prevCursor = potentialPrevCursor >= 0 ? potentialPrevCursor : undefined;
-
-    return {
-        ...(nextCursor === undefined ? {} : { next: { cursor: nextCursor.toString() } }),
-        ...(prevCursor === undefined ? {} : { prev: { cursor: prevCursor.toString() } }),
-    };
-};
 const CURRENCIES_WITH_APR = ['CAD'];
 
 export const calculateGrant = (amount: number | string, currency: string) => {
