+++ packages/shared/testing/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@integration-components/testing",
+    "version": "0.0.1",
+    "private": true,
+    "type": "module",
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "exports": {
+        ".": "./src/index.ts",
+        "./*": "./src/*"
+    }
+}
