+++ packages/shared/testing/src/index.ts
@@ -1,0 +1,2 @@
+export * as msw from './msw';
+export * as playwright from './playwright';
