+++ vite.config.ts
@@ -48,6 +48,7 @@ export default defineConfig(({ mode }) => {
                 '@integration-components/types': resolve(__dirname, 'packages/shared/types/src'),
                 '@integration-components/utils': resolve(__dirname, 'packages/shared/utils/src'),
                 '@integration-components/style': resolve(__dirname, 'packages/shared/style/src'),
+                '@integration-components/testing': resolve(__dirname, 'packages/shared/testing/src'),
             },
         },
         build: {
